+++ src/bin/spark.rs
@@ -1,24 +1,59 @@
+extern crate core;
+
 use std::io::{stdin, Read};
 
-use clap::{Arg, Command};
-use itertools::{Itertools, MinMaxResult};
+use clap::{arg, Arg, ArgGroup, Command};
+use itertools::Itertools;
 
 fn main() {
     let clap_config = Command::new("Spark")
         .about("Sparklines for the terminal")
         .version(env!("CARGO_PKG_VERSION"))
+        .arg(arg!(-t --ticks <TICKS> "The characters to use for the sparkline"))
+        .arg(arg!(--min <MIN> "The minimum value").value_parser(clap::value_parser!(f64)))
+        .arg(arg!(--max <MAX> "The maximum value").value_parser(clap::value_parser!(f64)))
         .arg(
             Arg::new("INPUT")
                 .help("The input to use, space or comma separated")
                 .allow_hyphen_values(true)
                 .num_args(0..)
                 .use_value_delimiter(true),
+        )
+        .group(
+            ArgGroup::new("min_max")
+                .args(["min", "max"])
+                .requires("min")
+                .requires("max")
+                .multiple(true),
         );
 
     let matches = clap_config.get_matches();
+    let default = sparklines::TICKS.as_slice();
+    let mut ticks = None;
+    if let Some(input_ticks) = matches.get_one::<String>("ticks") {
+        ticks = Some(input_ticks.chars().collect::<Vec<_>>());
+    }
+    if ticks.is_none() {
+        ticks = Some(default.to_vec());
+    }
+    let ticks = ticks.unwrap();
+
+    let mut min = None;
+    let mut max = None;
+    if let Some(input_min) = matches.get_one::<f64>("min") {
+        min = Some(*input_min);
+    }
+    if let Some(input_max) = matches.get_one::<f64>("max") {
+        max = Some(*input_max);
+    }
+    let string_spark = if let (Some(min), Some(max)) = (min, max) {
+        sparklines::StringSpark::new_with_min_max(&ticks, min, max)
+    } else {
+        sparklines::StringSpark::new(ticks.as_slice())
+    };
     if let Some(input) = matches.get_many::<String>("INPUT") {
         let numbers = input.flat_map(|s| s.parse::<f64>()).collect_vec();
-        println!("{}", spark(&numbers))
+        println!("{}", string_spark.spark(&numbers))
     } else {
         let mut input: Vec<u8> = Vec::new();
         if stdin().read_to_end(&mut input).is_ok() {
@@ -28,48 +63,8 @@ fn main() {
                     .flat_map(|s| s.split(','))
                     .flat_map(|v| v.parse::<f64>())
                     .collect::<Vec<_>>();
-                println!("{}", spark(&numbers));
+                println!("{}", string_spark.spark(&numbers));
             }
         }
-    }
-}
-
-const TICKS: [char; 8] = ['▁', '▂', '▃', '▄', '▅', '▆', '▇', '█'];
-
-fn spark(data: &[f64]) -> String {
-    let mut result = String::with_capacity(data.len() * 4);
-    let middle_idx = TICKS.len() / 2;
-    match data.iter().minmax() {
-        MinMaxResult::MinMax(min, max) => {
-            if min.eq(max) {
-                data.iter().for_each(|_| {
-                    result.push(TICKS[middle_idx]);
-                })
-            } else {
-                let f = (max - min) / (TICKS.len() - 1) as f64;
-                data.iter().for_each(|v| {
-                    let idx = ((v - min) / f) as usize;
-                    result.push(TICKS[idx]);
-                });
-            }
-        }
-        MinMaxResult::OneElement(_) => result.push(TICKS[middle_idx]),
-        MinMaxResult::NoElements => {}
-    }
-    result
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn graphs_data() {
-        assert_eq!(spark(&[1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0]), "▁▂▃▄▅▆▇█");
-    }
-
-    #[test]
-    fn equalizes_at_midtier_when_all_equal() {
-        assert_eq!(spark(&[1.0, 1.0, 1.0, 1.0]), "▅▅▅▅")
     }
 }